requirements.txt
asgiref==3.9.0
Django==5.2.4
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
mercadopago==2.2.1 